+++ blueprint/interfaces/billing/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @forklaunch/interfaces-billing
 
+## 1.0.25
+
+### Patch Changes
+
+- internal package upgrade
+
 ## 1.0.24
 
 ### Patch Changes